Clean up examples:
- remove rotten links
- rename folder for imperative examples
- move clear folder to example_dg_common.py
