+++ examples/dg/diffusion/example_dg_diffusion2D_qart.py
@@ -3,7 +3,6 @@ Based on
 
 Antonietti, P., & Quarteroni, A. (2013). Numerical performance of discontinuous and stabilized
     continuous Galerkin methods for convection–diffusion problems.
-Numerical Methods for Hyperbolic Equations, 75–85. https://doi.org/10.1201/b14172-9
 
 with advection part removed
 """
